pyproject.toml
[project]
name = "knowever-rss-digest"
version = "0.1.0"
description = "Prosty downloader RSS pod Codex – pobiera feedy do plików JSONL."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser",
    "PyYAML",
    "requests",
    "beautifulsoup4",
    "python-dotenv",
]

[project.scripts]
rss-download = "rss_download:download_all"
